Remove the legacy `URL` and `UrlDrop` classes
