+++ bridgetown-paginate/lib/bridgetown-paginate/pagination_page.rb
@@ -65,7 +65,9 @@ module Bridgetown
 
       def destination(dest)
         path = site.in_dest_dir(
-          dest, URL.unescape_path(url).delete_prefix(site.base_path(strip_slash_only: true))
+          dest, Bridgetown::Utils
+            .unencode_uri(url)
+            .delete_prefix(site.base_path(strip_slash_only: true))
         )
         path = File.join(path, "index") if url.end_with?("/")
         path << output_ext unless path.end_with? output_ext
